Partially import tables
Only inserts new records, doesn't do the merge correctly (yet)

No longer uses fastmember. Now it uses a custom dictionary data reader to load in a bunch of data to memory, and then use the bulk loader

There is a potentially better approach of making a reader that works directly with the XML file and streams into the bulk reader, making sure EnableStreaming is turned on

Another possible approach is to use DataTable which has Xml loading support built in. I'm not sure what the performance is for it.